chore(checkbox): remove unnecessary will-change (#2743)
Removes a couple of unnecessary usages of `will-change` in the checkbox. They are redundant, because the properties being animated aren't problematic to transition in the first place (`border-color`, `background-color` and `opacity`). In general, we should avoid adding `will-change` like this, until we actually notice a performance issue, and even in that case it should be added right before an animation is supposed to happen.