build: explicitly add tsc-wrapped to dev dependencies (#5029)
The dependency on `tsc-wrapped` should be added explicitly. Our tooling uses the `@angular/tsc-wrapped` package directly and it currently only works because the `@angular/compiler-cli` has a dependency on it.

If not specified explicitly, the `@angular/tsc-wrapped` dependency might be a local dependency of the compiler-cli and our tooling won't be able to access it anymore.