fix(menu): return focus to root trigger when closed by mouse (#8348)
Previously we only restored focus to the trigger if the user opened the menu using a keyboard, however that means that focus can be lost when closing by clicking on the backdrop. These changes switch to also restoring focus when clicking away, but only if the trigger is a top-level trigger.

Fixes #8290.