fix(drag-drop): item not being restored to its initial DOM position when using custom preview (#12808)
When a `cdkDrag` item is picked up, we save its `nextSibling` so we know where to put it once the user has stopped dragging. This logic gets thrown off when we use a custom preview element and it ends up leaving the element in the wrong place in the DOM. These changes ensure that the element is restored to its old position correctly.