fix(chips): chip list removing focus from first chip when adding through the input (#12840)
Fix regression in bcd0ed48c7ccfa90860454f42d62623150f0411d. Currently when the user removes all the chips and then they add a new chip, the chip list will remove focus from the input and put it on the chip. These changes introduce the proper behavior, which is to keep focus on the input.