fix(slide-toggle): thumb animation not working on mobile (#15236)
With https://github.com/angular/material2/commit/0a5057c82efa2fda4cb61bbcc584edf62f22bc5e, we enabled the `grabbing` cursor if the thumb container is pressed. This accidentally broke animation for the thumb sometimes because the `:active`
pseudo class is added on tap/touch and now accidentally sets the thumb transition duration to `0ms`. Rather we should only set the thumb transition duration to `0ms` if the user started dragging the thumb (with a small threshold)

Fixes #15232