build: do not run dgeni with  "info" log level. (#15512)
With Bazel it's uncommon to print anything except failures. Currently
the Dgeni logic that runs as part of our Bazel build is the only action
that prints something for a successful build. In order to make our Bazel
output clean, we lower the log level. For debugging, it's still easy to
modify the log-level.