fix(bottom-sheet): unable to reopen same bottom sheet after closing via back button (#15520)
Fixes not being able to reopen the same bottom sheet, if the user closed it by pressing back. The issue comes from the fact that we only consider the bottom sheet as closed when its animation completes, however that might not happen if the component is destroyed. The following changes work around the issue by completing the closing sequence on detach, rather than depending on the animation.

Fixes #15510.