fix(stepper): Fix visual flake in focusing stepper header (#15870)
When the step is changed through a button click sometimes the focus of the new
step header is attributed by focus monitor to 'mouse' instead of 'program'. This
results in a visual difference in the step header (focus background is missing).
Instead, when focusing the step header focus it via 'program' explicitly through
the focus monitor.